requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
requests==2.32.3
python-multipart==0.0.12
python-dotenv